+++ apps/appcontainer-node/packages/generic/src/appContainer.ts
@@ -34,6 +34,7 @@ import {
 	DataId,
 	LockId,
 	protectString,
+	getLogLevel,
 } from '@sofie-package-manager/api'
 
 import { WorkforceAPI } from './workforceApi'
@@ -275,6 +276,9 @@ export class AppContainer {
 	private async setupAvailableApps() {
 		const getWorkerArgs = (appId: AppId): string[] => {
 			return [
+				// Set initial loglevel to be same as appContainer:
+				`--logLevel=${getLogLevel()}`,
+
 				`--workerId=${appId}`,
 				`--workforceURL=${this.config.appContainer.workforceURL}`,
 				`--appContainerURL=${'ws://127.0.0.1:' + this.websocketServer?.port}`,
@@ -354,7 +358,7 @@ export class AppContainer {
 		// kill child processes
 	}
 	async setLogLevel(logLevel: LogLevel): Promise<void> {
-		this.logger.info(`Seting log level to "${logLevel}"`)
+		this.logger.info(`Setting log level to "${logLevel}"`)
 		setLogLevel(logLevel)
 	}
 	async _debugKill(): Promise<void> {