+++ shared/packages/api/src/logger.ts
@@ -43,6 +43,12 @@ export function initializeLogger(config: { process: ProcessConfig }): void {
 	process.on('warning', (e: any) => {
 		processLogger.error(`Unhandled warning: ${stringifyError(e)}`)
 	})
+
+	const initialLogLevel = config.process.logLevel
+	if (initialLogLevel && isLogLevel(initialLogLevel)) {
+		processLogger.info(`Setting log level to ${logLevel}`)
+		setLogLevel(initialLogLevel)
+	}
 }
 export function getLogLevel(): LogLevel {
 	if (!loggerContainer) throw new Error('Logging has not been set up! initializeLogger() must be called first.')
@@ -101,7 +107,9 @@ export function setupLogger(
 		})
 		if (initialLogLevel) setLogLevel(initialLogLevel, true)
 
-		logger.info('Logging to', logPath)
+		if (handleProcess) {
+			logger.info(`Logging to "${logPath}"`)
+		}
 	} else {
 		const transportConsole = new Winston.transports.Console({
 			level: logLevel,